FELIX-6746-websocketservlet-init-NPE - Remove util class, move method to handler - Add unit test, add test scope dependency on jetty12
