+++ http/base/pom.xml
@@ -157,5 +157,11 @@
             <version>5.7.0</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.eclipse.jetty.ee10.websocket</groupId>
+            <artifactId>jetty-ee10-websocket-jetty-server</artifactId>
+            <version>12.0.16</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
